Fix While Loop Paced Function Resolution
Always insert a paced promise resolve if a while look is present and the function already includes paced code.
Currently if an if statement is placed within a while loop with no other paced function call after the if statement, a browser crash will happen. This is due to the function promise resolution code "await new Promise(resolve => setTimeout(resolve," not being reachable unless the if statement does not resolve to true and thus the function does not resolve either.
Example code that will currently cause a crash is below.

The proposed code solution will work for any case where the function promise resolution cannot be reached inside the while loop.

Example Crash:

stage.whenFlag(function () {
    while(true) {
        if(stage.isKeyPressed('ArrowLeft')) {
            sprite.changeX(-20);
        }
    }
});